requirements.txt
python-telegram-bot==21.11.1
python-dotenv==1.0.1

